Start fleshing out partition change logic, reduce weaver id to int from long.

+++ spindle/src/main/java/com/salesforce/ouroboros/spindle/Bundle.java
@@ -49,5 +49,5 @@ public interface Bundle {
      * @param id
      * @param replicator
      */
-    void registerReplicator(long id, Replicator replicator);
+    void registerReplicator(int id, Replicator replicator);
 }
